Added notes to sunpath library
The current structure of the code is working fine but is not the best
option. I wanted to keep geometry libraries and calculation library
separate but now I think it's fine if geometry libraries has
dependencies on calculation libraries. This will make type checking much
easier as well as potentially make the same exact code to work in
different platforms. This is a change that I need to make soon before I
start any other dynamo libraries. Let's use sunpath as a test bed for
best practices in development for Dynamo.