Added documentation
After testing a couple of different python documentation tools I ended
up using pdoc. It's readable, visually appealing and includes the source
code in the documentation. I made a number of modifications to pdoc
(https://github.com/BurntSushi/pdoc/pull/95) and will make more changes
as the project goes forward. For example there is a need for providing a
link to higher level in documentation. Currently there is no easy way to
go up in the documentation.