PEP8 Corrections
I managed to run those packages eventuallly. So this file is checked for PEP8 for issues.

Now, speaking of test cases, the main() is not there in the file. Do you want me to create that function and put all the code inside it. I throught I may invite changes in other files that depend on this one so I thought I would ask first. Also, this will again raise line length related PEP8 issues.

For the time being, I commented out the test cases.

Pleaase let me know if I should put everything inside a main().

Also, In order to run this file to check everythng, I need to write
import __init__ as ladybug
or else it throws import errror. Just sharing. This is may be a gap in my understanding about the way core and [+] libraries are paired.