Cleaning Up WEA Class
Since all of the EPW-type sky models are living under this class, I am eventually planning on using it to help generate radiation values for AMY epw files that lack radiation data.  This commit involves some cleanup of the code for this purpose.  Namely, the radiation data lists now have setters, the timestep is a property that is linked with the data list, and the global horizontal radiation is a property.