Implemented Data Interpolation
The data interpolation now aligns with how E+ does it.

Notably, the EPW data now correctly aligns with its datetime in all cases as long as the data is not radiation or illuminance (it was originally off by an hour).  Furthermore the DataCollections all still count from the midnight before Jan 1 (1 Jan 00:00 ) so no functionality has been broken.  It is only the EPW data that has been shifted in order to correctly reflect datetimes.

The radiation and illuminance EPW data are still off by a half hour and, from what I can tell, simply changing the datetime to reflect this half-hour position should not break anything since this is how I am currently doing it with the ASHRAE Clear Skies and everything seems fine.  In fact, I am now concerned that HB+ annual simulations might be using the wrong direct sun positions since they might all be shifted by a half hour from the sun position that actually corresponds with the illuminance for that time.  @mostapharoudsari , we should discuss.

In addition to the interpolation, I fixed a bug in the datetime class, fixed PEP8 mistakes in DataCollection, and added a insert() method to the DataCollection class, which was needed to make this implementation possible.