Updated to Avoid Loss of Data
As soon as I made the commit last night, I regretted implementing it exactly as E+ does since this technically results in the last hour of the EPW being lost from the datastream.  With this implementation, I just move the last hour of the EPW into the 1 Jan 00:00 position, which preserves all the data in the epw object and results in all data (other then the half-hour data like radiation) being correctly aligned with its datetime.