Added Ability to Get Float Values from Data Collection
I have gotten so tired of putting [x.value for x in datacollection] in nearly all of my workflows with data collections.  The fact that we can't easily access the float values of the data collection has created a lot of difficulty and so I have added this property in for now.  I think that this is something that descriptors will eventually do a better job at fixing.