Discovered a Way that is More Elegant/Beautiful
Knowing that, in the future, we will certainly be extending the types namespace, I have modified the datatypes.py interface for the types namespace to dyanmically look up and read everything in the types folder.  This avoids having the data types hard coded into this interface, which I could see would definitely cause issues in the future.  Now, all that one has to do to add a new data type is either extend an existing base type or add in a new .py file for a new base type that follows the rules of the others.