Un-Hiding base module
I am realizing that people will want to access this module for any case where they are doing a from_json() and they want to write it in a way that works for all data types.  The litmus test for why I think it should be un-hidden is that I need to import it outside of the datatype module in the Header class.  If I can't use the datatype API properly without importing a hidden class, that indicates to me that it should be exposed.