Add Parsing of the EPW Header to EPW Object
Now that we have monthly collections, we have all of the objects that were needed to parse the header of the EPW into various other objects (including design days, analysis periods, and monthly collections).  Accordingly, this commit introduces parsing of the EPW header.  Editing any of these parsed objects will result in edits to the EPW header.