Added a Property for AnalysisPeriod Validation
I have come to realize that there is an important edge case that I originally overlooked, which I do not want to drive the development or increase the runtime of the current workflows but I think must nevertheless be addressed.  The edge case is if the user supplies datetimes that lie outside of the analysis_period in the header or supplies datetimes that are out of chronological order.  I say that this is an edge case because there is no need for these checks whenever the data is derived from a continuous data set, which is pretty much 99% of Ladybug workflows.  Furthermore, the vast majority of the methods on the DataCollection will still work without these checks and I only see it becoming an issue for certain types of visualizations or when one wants to build a continuous collection from a discontinuous one. Accordingly, to solve this issue, I have added an attribute to the DataCollections to track whether the header analysis_period aligns with the datetimes in the data, which is always True through all operations on a continuous collection.  I have also added methods that validate the datetimes against the analysis_period.