Small Change to datatype
This includes a few more changes to the datatype class in order for it to be able to handle thermal comfort conditions.  Specifically, I have moved the ThermalComfort datatype under the thermalcondition module where it seems to fit better.  I am also replacing the convention of unit_descr as a string with one as a dictionary.  This will make it easier to use this unit description to build ordinal legends when we start integrating datatype with legendpar.