Consolidated Code
There was a lot of duplicated code in the DataCollection methods that computed averaged_monthy, total_monthly, averaged_daily, etc.  These have now all been consolidated into a single hidden function with arguments for the time interval and the operation to perform.  In the process, I have gotten rid of 100+ lines of duplicated code and made it easy to implement monthly operations on te DailyCollection.