feat(datatype): Add Degree-Days and Degree-Hours
I also really started regretting that the data type for handling fractional values was written as 'percentage' instead of 'fraction.'  It seems that I was narrow-mindedly thinking of relative humidity in EPW files when I set up that type.  So I am fixing it here before we start having more dependencies on it.  Right now, it shouldn't really break anything outdside of the API.