fix(deploy): use javascript semantic-release package for deployment
Using the "pure" javascript implementation of semantic-release makes it a bit more flexible to set
up custom scripts to run at different stages of the semantic release process. The packages used as
well as scripts to run at different stages are specified in the .releaserc.json file. In this case
we use standard semantic-release and add a custom deploy script on top of the normal github release
setup. The custom deploy script is the new 'deploy.sh' file which takes the new version number as an
argument to deploy the package to PyPi and build the documentation.

#125

build(pytest): add pytest-cov deps to dev-requirements.txt

build(euclid): add euclid to dev-requirements.txt