feat(date, time): Add objects for Date and Time
These two objects are particularly helpful for defining EnergyPlus schedules and there may be some other places where it is helpful to have these objects coordinated with the Ladybug DateTime and AnalysisPeriod as they are here.

This commit also adds a full set of tests for DateTime, Date, and Time.  It also includes a few corrections to places where we were internally using hasattr instead of isinstance.