feat(futil): Remove bat_to_sh
As per our recent discussion, @mostapharoudsari , I want to remove this bat_to_sh function before any other parts of Ladybug Tools make use of it (right now, there's nothing using it and it's only here because it was originally a part of the honeybee.fuitl module that this was derived from).

Lastly, there are some docstring improvements.