feat(logger): Add logger
Added a logger that should be used from now on instead of using print. I have also made sure that the folder in which the logger is logging matches that of the default_epw_folder in the config.

One question is if we should update honeybee-core to use some of the elements of this logger and, if so, what elements should these be? Should it just be importing ladybug's logutil and setting it to write to honeybee.log instead of ladybug.log or is there a deeper change that needs to happen.