fix(designday): Refactor the DesignDay class with best practices
This commit is trying to nip in the bud what I know will be a major naming issue in the future. Right now, the DesignDay object has properties for `ep_style_string`, which manages to be both verbose and yet also vague. Particularly if/when the epjson format replaces the idf format, it will be tough to know whether the "ep_style" refers to an idf or epjson. So I'm replacing all instances of this with `to_idf()` and `from_idf()` as we have it in honeybee_energy.

I'm also implementing some better practices taken from `honeybee_energy`, including the adding of `duplicate` methods along with `==`, `!=`, and `hash` methods.

Lastly, I am trying to use better word choice for the dict keys of the design day. knowing that this is my last chance to change them without breaking `honeybee_energy` or pydantic schema documentation.