fix(designday): Use Date object for DesignDay
I realized that the only reason why I was requesting separate inputs for month and day_of_month was because I originally wrote this class before we had Date objects in the dt module. So this commit is replacing these two inputs with a Date object and this should be less prone to errors (like invalid dates).