fix(wea): Simplify Wea dictionary schema
There were several redundant properties within the Wea schema, which was fine when all dictionaries were originating from the ladybug-core library. But, now that I have been documenting the Wea schema for radiance workflows, the schema really should be refactored to minimize the potential for conflicts and need for checks between these redundant properties.

I also added a check that the location assigned to the Wea is a Location object.