feat(chart): Add MonthlyChart and fix several bugs
This commit fixes a few issues with the data collection and data type classes.

It also adds an option into the hourly plot to reverse the Y-axis.

Lastly, it adds a class for creating monthly charts from any data collection. The default ranges it uses for the axes are MUCH better than legacy given that it uses the cumulative property on the data type to decide whether the plot should start from 0.