perf(analysisperiod) Improve performance of AnalysisPeriod datetimes
This commit includes a number of performance improvements to the analysis period and (by extension) the hourly continuous data collection.  The biggest improvement comes from the fact that we can get the length of the analysis period without having to create all of the individual datetime objects (we just look at the time difference between the start and end datetimes). Also, when a case arrives that datetimes are needed, these datetimes get cached on the AnalysisPeriod object instead of regenerating them every time that they are called. This should dramatically cut down on the number of datetime objects that we are storing in memory at a given time.

The improvement is good enough that I can even notice the tests running faster.

Technically, there's also a bug fix in this commit but it affects such a niche edge case (reversed analysis periods for leap years) that I don't think it's worth a whole new release right now.