feat(epw): Add method to derive DDY from EPW
This commit adds a method to the EPW class that can derive a DDY from the EPW file. The DDYs produced by this method are not as representative of the climate as the ones distributed with the EPWs but this will help a lot when running energy simulations with EPW files that have no design associated with them.

Resolves https://github.com/ladybug-tools/ladybug/issues/308