feat(wea): Add support for discontinuous data in Wea objects
This commit makes it possible to use hourly discontinuous data collections for Wea objects. It also adds a range of methods to produce filtered Weas based on certain criteria including:

* filter_by_pattern
* filter_by_analysis_period
* filter_by_hoys
* filter_by_moys
* filter_by_sun_up

I also fixed a bug in the previous performance improvement of AnalysisPeriod.__len__() that I added last week. And I added a from_start_end_datetime method to AnalysisPeriod that I am already making use of in this commit and I know I will use elsewhere. Lastly, I verified that the AnalysisPeriod.from_analysis_period method is not being used by any of the Grasshopper plugins and I removed it before any other developers feel the temptation to make use of it. It just feels like it's a liability the longer it remains in the code base and it's best to rip the band aid off now.

Lastly, I fixed a small bug in the data collection __key() property and improved / standardized a few of the docstrings that I came across.

There are a couple of Grasshopper components using the Wea class that I should update to align with this PR but, otherwise, this is all ready to merge.

Resolves https://github.com/ladybug-tools/ladybug/issues/265