feat(cli): Add a CLI and basic translation commands
I am also refactoring all of the classes that can write to files to have a write() method for consistency (we will leave the save() method on EPW and DDY for the time being).

The code that actually generates the file contents is now its own method so we can always get the string if there's something else handling the writing-to-file (eg. a CLI that may want to print things to stdout).

Lastly, I added an EPW.from_file_string, which we can use for things like a web interface for ladybug.