fix(cli): Replace hacky use of analysis period JSON with string
I realized that the AnalysisPeriod object already has all of its properties within its string representation. So just having people pass a string of an analysis period to the CLI will be much cleaner. And it allows us to get around the fact that queenbee does not support conditional artifacts.